dashboard/requirements.txt
# Dashboard + Runtime dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
psycopg2-binary>=2.9.0    # PostgreSQL
PyJWT>=2.8.0               # JWT tokens
google-auth>=2.20.0        # Google OAuth token verification
requests>=2.28.0           # Required by google-auth transport
python-dotenv>=1.0.0       # Auto-load .env file
pytest>=7.0.0              # Test runner (used by agent-built tests)
httpx>=0.24.0              # Required by FastAPI TestClient (starlette dependency)
